refactor(constants): migrate enums from types/enums.ts to constant.ts for better organization and consistency
