+++ src/handlers/index.ts
@@ -1,8 +1,3 @@
-/**
- * 导出所有 command handlers
- * 这些 handlers 负责处理 VSCode 命令和编辑器事件
- */
-
 export { createConvertHandler } from './convertHandler';
 export { createPasteHandler } from './pasteHandler';
 export { createUndoHandler } from './undoHandler';
