+++ src/i18n.ts
@@ -4,7 +4,8 @@ import Hook from './hook';
 import { getConfig } from "./config";
 import { FILE_IGNORE } from './constant';
 import { getWorkspaceKey } from './utils';
-import Watcher, { WATCH_STATE } from './watcher';
+import Watcher from './watcher';
+import { WatchState } from './constant';
 
 import type { Uri } from 'vscode';
 import type { I18nGroup } from "./types";
@@ -59,16 +60,16 @@ export default class I18n {
             return;
         }
 
-        const watchCallback = async (state: WATCH_STATE, uri: Uri) => {
+        const watchCallback = async (state: WatchState, uri: Uri) => {
             const pathMap = this.i18nMap.get(workspaceKey) || new Map();
 
             switch (state) {
-                case WATCH_STATE.CREATE:
-                case WATCH_STATE.CHANGE:
+                case WatchState.Create:
+                case WatchState.Change:
                     pathMap.set(uri.fsPath, await Hook.getInstance().collectI18n({ i18nFileUri: uri }));
                     this.i18nMap.set(workspaceKey, pathMap);
                     break;
-                case WATCH_STATE.DELETE:
+                case WatchState.Delete:
                     pathMap.delete(uri.fsPath);
                     this.i18nMap.set(workspaceKey, pathMap);
                     break; 
@@ -80,7 +81,7 @@ export default class I18n {
         // init
         const i18nFileUris = await workspace.findFiles(i18nFilePattern, FILE_IGNORE);
         for (const uri of i18nFileUris) {
-            await watchCallback(WATCH_STATE.CHANGE, uri);
+            await watchCallback(WatchState.Change, uri);
         }
 
         const watcher = await new Watcher().watch(i18nFilePattern, watchCallback);
