+++ src/types/index.ts
@@ -1,6 +1,6 @@
 import type { Range, DecorationOptions, Definition, DefinitionLink } from 'vscode';
 import type { MessageFormatElement } from '@formatjs/icu-messageformat-parser';
-import type { SupportType, ConvertType, ConflictPolicy } from './enums';
+import type { SupportType, ConvertType, ConflictPolicy } from '../constant';
 
 export type ConvertGroup = {
     i18nValue: string;
