+++ src/watcher.ts
@@ -4,14 +4,10 @@ import { workspace, Uri, RelativePattern } from 'vscode';
 import { watch } from 'chokidar';
 import { isString } from 'lodash';
 
+import { WatchState } from './constant';
+
 import type { GlobPattern, Disposable } from "vscode";
 import type { FSWatcher } from 'chokidar';
-
-export enum WATCH_STATE {
-    CHANGE = 'change',
-    CREATE = 'create',
-    DELETE = 'delete',
-}
 
 export default class Watcher implements Disposable {
     private cwd: string;
@@ -35,7 +31,7 @@ export default class Watcher implements Disposable {
         return pattern;
     }
 
-    async watch(pattern: GlobPattern, callback: (state: WATCH_STATE, uri: Uri) => void): Promise<Watcher> {
+    async watch(pattern: GlobPattern, callback: (state: WatchState, uri: Uri) => void): Promise<Watcher> {
         const rp = this.toRelativePattern(pattern);
         const cwd = rp.baseUri?.fsPath || this.cwd;
 
@@ -47,9 +43,9 @@ export default class Watcher implements Disposable {
             });
 
             this.watcher
-                .on('add', (relativePath) => callback(WATCH_STATE.CREATE, Uri.file(path.join(cwd, relativePath))))
-                .on('change', (relativePath) => callback(WATCH_STATE.CHANGE, Uri.file(path.join(cwd, relativePath))))
-                .on('unlink', (relativePath) => callback(WATCH_STATE.DELETE, Uri.file(path.join(cwd, relativePath))))
+                .on('add', (relativePath) => callback(WatchState.Create, Uri.file(path.join(cwd, relativePath))))
+                .on('change', (relativePath) => callback(WatchState.Change, Uri.file(path.join(cwd, relativePath))))
+                .on('unlink', (relativePath) => callback(WatchState.Delete, Uri.file(path.join(cwd, relativePath))))
                 .on('error', (error) => console.error(`Watcher(cwd: ${cwd}, pattern: ${rp.pattern}) error:`, error))
                 .on('ready', () => resolve(this));
         });
