+++ test/react/.vscode/i18n-fast.hook.js
@@ -13,11 +13,11 @@ const genI18nKey = require('./gen-i18n-key');
 */
 
 /**
- * @typedef {'document'} MatchType - see {@link https://github.com/lvboda/vscode-i18n-fast/blob/main/src/types/enums.ts}
+ * @typedef {'document'} MatchType - see {@link https://github.com/lvboda/vscode-i18n-fast/blob/main/src/constant.ts}
 */
 
 /**
- * @typedef {0 | 1 | 2 | 4 | 7} SupportType - see {@link https://github.com/lvboda/vscode-i18n-fast/blob/main/src/types/enums.ts}
+ * @typedef {0 | 1 | 2 | 4 | 7} SupportType - see {@link https://github.com/lvboda/vscode-i18n-fast/blob/main/src/constant.ts}
 */
 
 /**
